Migrate user assets to `createQueryStore`
