--- src/__swaps__/screens/Swap/resources/assets/index.ts
@@ -1,3 +1,0 @@
-export { useUserAssets } from './userAssets';
-export type { UserAssetsArgs } from './userAssets';
-export type { UserAssetsByChainArgs } from './userAssetsByChain';
