+++ src/hooks/useAccountSettings.ts
@@ -1,3 +1,4 @@
+import { Address } from 'viem';
 import lang from 'i18n-js';
 import { useCallback } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
@@ -12,6 +13,8 @@ import {
 import { AppState } from '@/redux/store';
 import { supportedNativeCurrencies } from '@/references';
 import { NativeCurrencyKey } from '@/entities';
+import { userAssetsStoreManager } from '@/state/assets/userAssetsStoreManager';
+import { useConnectedToAnvilStore } from '@/state/connectedToAnvil';
 
 const languageSelector = (state: AppState) => state.settings.language;
 
@@ -22,23 +25,26 @@ const withLanguage = (language: string) => {
   return { language };
 };
 
+const FALLBACK_ADDRESS = userAssetsStoreManager.getState().address;
 const createLanguageSelector = createSelector([languageSelector], withLanguage);
 
 export default function useAccountSettings() {
   const { language } = useSelector(createLanguageSelector);
   const dispatch = useDispatch();
-  const settingsData = useSelector(
-    ({ settings: { accountAddress, appIcon, chainId, nativeCurrency, network, testnetsEnabled } }: AppState) => ({
-      accountAddress,
-      appIcon,
-      chainId,
-      language,
-      nativeCurrency,
-      nativeCurrencySymbol: supportedNativeCurrencies[nativeCurrency as keyof typeof supportedNativeCurrencies].symbol,
-      network,
-      testnetsEnabled,
-    })
-  );
+
+  const nativeCurrency = userAssetsStoreManager(state => state.currency);
+  const testnetsEnabled = useConnectedToAnvilStore(state => state.connectedToAnvil);
+
+  const settingsData = useSelector(({ settings: { accountAddress, appIcon, chainId, network } }: AppState) => ({
+    accountAddress: (accountAddress.length ? accountAddress : FALLBACK_ADDRESS ?? accountAddress) as Address,
+    appIcon,
+    chainId,
+    language,
+    nativeCurrency,
+    nativeCurrencySymbol: supportedNativeCurrencies[nativeCurrency].symbol,
+    network,
+    testnetsEnabled,
+  }));
 
   const settingsChangeLanguage = useCallback((language: string) => dispatch(changeLanguage(language as Language)), [dispatch]);
