+++ src/screens/SendSheet.tsx
@@ -129,7 +129,7 @@ type OnSubmitProps = {
 export default function SendSheet() {
   const { goBack, navigate } = useNavigation();
   const sortedAssets = useUserAssetsStore(state => state.legacyUserAssets);
-  const isLoadingUserAssets = useUserAssetsStore(state => state.isLoadingUserAssets);
+  const isLoadingUserAssets = useUserAssetsStore(state => state.getStatus().isInitialLoading);
   const {
     gasFeeParamsBySpeed,
     gasLimit,
